birthweight on stunting bugfix (#5)
    Category: bugfix
    JIRA issue: MIC-4387

Changes and notes
There was a silent division by 0 error leading to NaNs in our pipelines when we had stunting exposure probabilities of cat1 and cat2 summing to 0. This case should only happen when all the probability of stunting exposure is allocated to cat4. In this case, there is no effect of birthweight on stunting exposure.

Verification and Testing
Initialized population and checked that everyone was unexposed.