add all transitions to PEM component (#10)
    Category: bugfix
    JIRA issue: MIC-4403

Changes and notes
In a RiskAttributableDisease, the infected to susceptible transition is added to the transition names attribute on setup. The DiseaseObserver registers these transition observations in its setup which occurs before PEM's so it never gets registered. Add this transition on instantiation instead.

Verification and Testing
Ran a sim for a few steps and saw these observation transitions in the output.